+++ runtime/execdriver/lxc/driver.go
@@ -2,7 +2,7 @@ package lxc
 
 import (
 	"fmt"
-	"github.com/dotcloud/docker/execdriver"
+	"github.com/dotcloud/docker/runtime/execdriver"
 	"github.com/dotcloud/docker/pkg/cgroups"
 	"github.com/dotcloud/docker/utils"
 	"io/ioutil"
