+++ runtime/execdriver/native/term.go
@@ -5,7 +5,7 @@
 package native
 
 import (
-	"github.com/dotcloud/docker/execdriver"
+	"github.com/dotcloud/docker/runtime/execdriver"
 	"io"
 	"os"
 	"os/exec"
